refactor(fabric-tests): consolidate duplicate test infrastructure (~1200 lines removed)
  Merge 3 multicast conn_mgr kernels (basic/scatter/fused) into 1 unified
  kernel dispatching on OPERATION_TYPE via constexpr-if. Merge unicast and
  multicast auto-packetization kernels into tx_writer_raw.cpp via
  IS_MULTICAST define. Consolidate 3 addrgen runners and 2 auto-pack
  runners into single runner.cpp files. Extract shared test helpers into
  runner_common.hpp. Remove 4 redundant compile-only probe kernels and
  their test classes.

  Fix NocUnicastScatterCommandHeader construction across mesh/api.h,
  linear/api.h, and test kernels — wrap bare uint16_t chunk_size args
  in initializer-list braces to match constructor signature.
