--- tests/tt_metal/tt_fabric/fabric_data_movement/auto_packetization/kernels/linear_unicast_tx_writer_raw.cpp
@@ -1,45 +1,0 @@
-// SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
-//
-// SPDX-License-Identifier: Apache-2.0
-
-// 1D (linear) compile probe kernel for auto-packetization.
-// Only includes linear/api.h — does NOT require FABRIC_2D.
-
-#include <cstdint>
-#include "api/dataflow/dataflow_api.h"
-#include "fabric/fabric_edm_packet_header.hpp"
-#include "tt_metal/fabric/hw/inc/edm_fabric/edm_fabric_worker_adapters.hpp"
-#include "tt_metal/fabric/hw/inc/packet_header_pool.h"
-#include "tt_metal/fabric/hw/inc/tt_fabric_api.h"
-#include "tt_metal/fabric/hw/inc/noc_addr.h"
-#include "tt_metal/fabric/hw/inc/linear/api.h"
-
-using namespace tt::tt_fabric;
-using namespace tt::tt_fabric::linear::experimental;
-
-void kernel_main() {
-    size_t idx = 0;
-    const uint32_t src_l1_addr   = get_arg_val<uint32_t>(idx++);
-    const uint32_t total_size    = get_arg_val<uint32_t>(idx++);
-    const uint32_t dst_base_addr = get_arg_val<uint32_t>(idx++);
-    const uint32_t rx_noc_x      = get_arg_val<uint32_t>(idx++);
-    const uint32_t rx_noc_y      = get_arg_val<uint32_t>(idx++);
-    const uint8_t  num_hops      = static_cast<uint8_t>(get_arg_val<uint32_t>(idx++));
-
-    auto sender = WorkerToFabricEdmSender::build_from_args<ProgrammableCoreType::TENSIX>(idx);
-    volatile tt_l1_ptr PACKET_HEADER_TYPE* packet_header = PacketHeaderPool::allocate_header();
-    sender.open<true>();
-
-    const uint64_t dst_noc_addr = safe_get_noc_addr(rx_noc_x, rx_noc_y, dst_base_addr, 0);
-
-    // Auto-packetizing linear unicast write
-    fabric_unicast_noc_unicast_write(
-        &sender,
-        packet_header,
-        src_l1_addr,
-        total_size,
-        tt::tt_fabric::NocUnicastCommandHeader{dst_noc_addr},
-        num_hops);
-
-    sender.close();
-}
