+++ tests/tt_metal/tt_fabric/fabric_data_movement/auto_packetization/kernels/tx_writer_raw.cpp
@@ -2,14 +2,21 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-// Unified multicast TX kernel for auto-packetization integration tests.
-// Dispatches to the correct fabric multicast API based on the send_op RT arg.
+// Unified TX kernel for auto-packetization integration tests.
+// Handles both unicast and multicast via IS_MULTICAST compile-time define.
+// Dispatches to the correct fabric API based on the send_op RT arg.
 //
-// RT args:
+// Unicast RT args:
+//   0: src_l1_addr, 1: total_size, 2: dst_base_addr,
+//   [FABRIC_2D] 3: dst_mesh_id, 4: dst_dev_id
+//   next: rx_noc_x, rx_noc_y, sem_l1_addr
+//   [!FABRIC_2D] next: num_hops
+//   next: scatter_offset, send_op, ...fabric connection args
+//
+// Multicast RT args:
 //   0: src_l1_addr, 1: total_size, 2: dst_base_addr,
 //   3: rx_noc_x, 4: rx_noc_y, 5: sem_l1_addr
-//   6: scatter_offset (u32, 0 for non-scatter ops)
-//   7: send_op        (u32, one of TX_OP_*)
+//   6: scatter_offset, 7: send_op
 //   [FABRIC_2D] next: dir_mask, senderW/E/N/S args, e/w/n/s_hops
 //   [!FABRIC_2D] next: start_distance, range, sender args
 
@@ -17,6 +24,9 @@
 
 void kernel_main() {
     size_t idx = 0;
+
+#ifdef IS_MULTICAST
+    // ===== MULTICAST MODE =====
     const uint32_t src_l1_addr   = get_arg_val<uint32_t>(idx++);
     const uint32_t total_size    = get_arg_val<uint32_t>(idx++);
     const uint32_t dst_base_addr = get_arg_val<uint32_t>(idx++);
@@ -158,4 +168,92 @@ void kernel_main() {
 
     sender.close();
 #endif
+
+#else
+    // ===== UNICAST MODE =====
+    TX_KERNEL_PARSE_UNICAST_ARGS(idx)
+    const uint32_t scatter_offset = get_arg_val<uint32_t>(idx++);
+    const uint32_t send_op        = get_arg_val<uint32_t>(idx++);
+    TX_KERNEL_SETUP(idx)
+
+    const uint64_t dst_noc_addr  = safe_get_noc_addr(rx_noc_x, rx_noc_y, dst_base_addr, 0);
+    const uint64_t dst_noc_addr1 = safe_get_noc_addr(rx_noc_x, rx_noc_y, dst_base_addr + scatter_offset, 0);
+    const uint64_t sem_noc_addr  = safe_get_noc_addr(rx_noc_x, rx_noc_y, sem_l1_addr, 0);
+    const uint16_t scatter_chunk_size = static_cast<uint16_t>(total_size / 2);
+
+    if (send_op == TX_OP_WRITE) {
+#ifdef FABRIC_2D
+        fabric_unicast_noc_unicast_write(
+            &sender, packet_header, dst_dev_id, dst_mesh_id,
+            src_l1_addr, total_size, tt::tt_fabric::NocUnicastCommandHeader{dst_noc_addr});
+#else
+        fabric_unicast_noc_unicast_write(
+            &sender, packet_header,
+            src_l1_addr, total_size, tt::tt_fabric::NocUnicastCommandHeader{dst_noc_addr}, num_hops);
+#endif
+        noc_async_writes_flushed();
+#ifdef FABRIC_2D
+        fabric_unicast_noc_unicast_atomic_inc(
+            &sender, packet_header, dst_dev_id, dst_mesh_id,
+            tt::tt_fabric::NocUnicastAtomicIncCommandHeader(sem_noc_addr, 1, 32));
+#else
+        fabric_unicast_noc_unicast_atomic_inc(
+            &sender, packet_header,
+            tt::tt_fabric::NocUnicastAtomicIncCommandHeader(sem_noc_addr, 1, 32), num_hops);
+#endif
+    } else if (send_op == TX_OP_SCATTER_WRITE) {
+#ifdef FABRIC_2D
+        fabric_unicast_noc_scatter_write(
+            &sender, packet_header, dst_dev_id, dst_mesh_id,
+            src_l1_addr, total_size,
+            tt::tt_fabric::NocUnicastScatterCommandHeader{{dst_noc_addr, dst_noc_addr1}, {scatter_chunk_size}});
+#else
+        fabric_unicast_noc_scatter_write(
+            &sender, packet_header,
+            src_l1_addr, total_size,
+            tt::tt_fabric::NocUnicastScatterCommandHeader{{dst_noc_addr, dst_noc_addr1}, {scatter_chunk_size}},
+            num_hops);
+#endif
+        noc_async_writes_flushed();
+#ifdef FABRIC_2D
+        fabric_unicast_noc_unicast_atomic_inc(
+            &sender, packet_header, dst_dev_id, dst_mesh_id,
+            tt::tt_fabric::NocUnicastAtomicIncCommandHeader(sem_noc_addr, 1, 32));
+#else
+        fabric_unicast_noc_unicast_atomic_inc(
+            &sender, packet_header,
+            tt::tt_fabric::NocUnicastAtomicIncCommandHeader(sem_noc_addr, 1, 32), num_hops);
+#endif
+    } else if (send_op == TX_OP_FUSED_ATOMIC_INC) {
+#ifdef FABRIC_2D
+        fabric_unicast_noc_fused_unicast_with_atomic_inc(
+            &sender, packet_header, dst_dev_id, dst_mesh_id,
+            src_l1_addr, total_size,
+            tt::tt_fabric::NocUnicastAtomicIncFusedCommandHeader{dst_noc_addr, sem_noc_addr, 1, true});
+#else
+        fabric_unicast_noc_fused_unicast_with_atomic_inc(
+            &sender, packet_header,
+            src_l1_addr, total_size,
+            tt::tt_fabric::NocUnicastAtomicIncFusedCommandHeader{dst_noc_addr, sem_noc_addr, 1, true},
+            num_hops);
+#endif
+    } else {  // TX_OP_FUSED_SCATTER_ATOMIC_INC
+#ifdef FABRIC_2D
+        fabric_unicast_noc_fused_scatter_write_atomic_inc(
+            &sender, packet_header, dst_dev_id, dst_mesh_id,
+            src_l1_addr, total_size,
+            tt::tt_fabric::NocUnicastScatterAtomicIncFusedCommandHeader{
+                {dst_noc_addr, dst_noc_addr1}, sem_noc_addr, {scatter_chunk_size}, 1, true});
+#else
+        fabric_unicast_noc_fused_scatter_write_atomic_inc(
+            &sender, packet_header,
+            src_l1_addr, total_size,
+            tt::tt_fabric::NocUnicastScatterAtomicIncFusedCommandHeader{
+                {dst_noc_addr, dst_noc_addr1}, sem_noc_addr, {scatter_chunk_size}, 1, true},
+            num_hops);
+#endif
+    }
+
+    TX_KERNEL_TEARDOWN()
+#endif  // IS_MULTICAST
 }
