+++ tests/tt_metal/tt_fabric/sources.cmake
@@ -31,12 +31,9 @@ set(UNIT_TESTS_FABRIC_SRC
     fabric_data_movement/test_basic_fabric_mux.cpp
     fabric_data_movement/test_fabric_traffic_generator_kernel.cpp
     fabric_data_movement/test_auto_packetization.cpp
-    fabric_data_movement/auto_packetization/unicast_runner.cpp
-    fabric_data_movement/auto_packetization/multicast_runner.cpp
+    fabric_data_movement/auto_packetization/runner.cpp
     fabric_data_movement/addrgen_write/test_addrgen_write.cpp
-    fabric_data_movement/addrgen_write/unicast_runner.cpp
-    fabric_data_movement/addrgen_write/multicast_runner.cpp
-    fabric_data_movement/addrgen_write/linear_multicast_runner.cpp
+    fabric_data_movement/addrgen_write/runner.cpp
 )
 
 set(UNIT_TESTS_PHYSICAL_DISCOVERY_SRC physical_discovery/test_physical_system_descriptor.cpp)
